// src/api/__generated__/story/model/boundActorCollectionResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */

export interface BoundActorCollectionResponse {
  actorCollectionId?: number;
  assetId?: string;
  name?: string;
  avatarUrl?: string;
  trust?: number;
  computingPower?: number;
}